+++ sampledb/logic/objects.py
@@ -16,7 +16,7 @@ import typing
 from ..models import Objects, Object, Action, ActionType
 from . import object_log, user_log, object_permissions, errors, users, actions, tags
 from .notifications import create_notification_for_being_referenced_by_object_metadata
-from .errors import CreatingObjectsDisabled
+from .errors import CreatingObjectsDisabledError
 import sqlalchemy.exc
 
 
@@ -53,9 +53,8 @@ def create_object(
         user ID exists
     """
     action = actions.get_action(action_id)
-    action_type = actions.get_action_type(action.type_id)
-    if action_type.disable_create_objects:
-        raise CreatingObjectsDisabled
+    if action.type.disable_create_objects:
+        raise CreatingObjectsDisabledError()
     users.get_user(user_id)
     try:
         object = Objects.create_object(data=data, schema=schema, user_id=user_id, action_id=action_id)