+++ sampledb/static/js/schema_editor.js
@@ -6,6 +6,21 @@ $(function() {
   input_schema.hide();
   $('.pygments').hide();
 
+  // Add function to hide alert onclick
+  $('#schema-editor-type-is-template-close').click(function() {
+      $('.schema-editor-type-is-template').hide();
+  });
+  // Add function to show alert if action_type data-is-template is True on change
+  let type_input =  $('#input-type');
+  type_input.change(function() {
+      let chosen_action_type = type_input.find('[value=' + type_input.val() + ']');
+      if(chosen_action_type.data('isTemplate') === "True") {
+          $('.schema-editor-type-is-template').show();
+      } else {
+          $('.schema-editor-type-is-template').hide();
+      }
+  }).change();
+
   var schema = JSON.parse(input_schema.text());
   delete schema['displayProperties'];
   if ('propertyOrder' in schema && schema['propertyOrder'].includes('tags')) {
@@ -193,20 +208,22 @@ $(function() {
         var is_required = ('required' in schema && schema['required'].includes(name));
         var property_node = buildPropertyNode(schema['properties'][name], path, is_required);
 
-        if (property_node !== null && 'template' in schema['properties'][name]) {
-          var select_input = $(property_node[0]).find('#schema-editor-object__' + name + '-template-id-input')[0];
-          var act_template = schema['properties'][name]['template'];
-          $(select_input.options).each(function() {
-            if(this.value == act_template) {
-              $(this).attr('selected', true);
-            } else {
-              $(this).attr('selected', false);
-            }
-          });
-          $(select_input).selectpicker('refresh');
-          properties_node[0].appendChild(property_node[0]);
-        } else if (property_node !== null) {
-          properties_node[0].appendChild(property_node[0]);
+        if (property_node !== null) {
+          if ('template' in schema['properties'][name]) {
+            var select_input = $(property_node[0]).find('#schema-editor-object__' + name + '-template-id-input')[0];
+            var current_template = schema['properties'][name]['template'];
+            $(select_input.options).each(function() {
+              if (Number.parseInt(this.value) === current_template) {
+                $(this).attr('selected', true);
+              } else {
+                $(this).attr('selected', false);
+              }
+            });
+            $(select_input).selectpicker('refresh');
+            properties_node[0].appendChild(property_node[0]);
+          } else {
+            properties_node[0].appendChild(property_node[0]);
+          }
         }
       }
     }
@@ -300,7 +317,7 @@ $(function() {
     $('button[name="action_submit"]').prop('disabled', (JSON.stringify(window.schema_editor_errors) !== '{}'));
   }
 
-  function buildPropertyNode(schema, path, is_required, template_id = null) {
+  function buildPropertyNode(schema, path, is_required) {
     var type;
     if (schema['type'] === 'text') {
       if ('choices' in schema) {
@@ -372,8 +389,8 @@ $(function() {
         updateDatetimeProperty(path, real_path);
       } else if (type === "plotly_chart") {
         updatePlotlyChartProperty(path, real_path);
-      } else if(type === 'template') {
-        updateTemplateObjectProperty(path, real_path, template_id);
+      } else if (type === 'template') {
+        updateTemplateObjectProperty(path, real_path);
       }
       globallyValidateSchema();
     }
@@ -492,15 +509,11 @@ $(function() {
       var node = $(this).closest('.schema-editor-generic-property');
       node.find('.schema-editor-property-settings').css('display', 'none');
       $(node.find('.schema-editor-property-settings')[0]).css('display', 'flex');
+      node.find('.help-block').html('');
+      node.find('.has-error').removeClass('has-error');
       node.find('.schema-editor-' + type + '-property-settings').css('display', 'flex');
     });
     type_input.on('change', updateProperty.bind(path));
-
-    var template_id_input = node.find('.schema-editor-generic-property-template-id-input');
-    template_id_input.attr('id', 'schema-editor-object__' + path.join('__') + '-template-id-input');
-    template_id_input.val(type);
-    template_id_input.selectpicker();
-    template_id_input.on('change', updateProperty.bind(path));
 
     function setOptionalValueInPropertySchema(path, type, name, property_schema) {
       var has_value = $('#schema-editor-object__' + path.join('__') + '-' + type + '-' + name + '-checkbox').prop('checked');
@@ -630,19 +643,6 @@ $(function() {
       } else {
         title_help.text("");
         title_group.removeClass("has-error");
-      }
-      var template_input = $('#schema-editor-object__' + path.join('__') + '-template-id-input');
-      var template_group = template_input.closest('.schema-editor-property-settings.schema-editor-template-property-settings');
-      var template_help = template_group.find('.help-block');
-      if($(template_input).is(':visible')) {
-        if($( template_input ).val() === "") {
-          template_help.text(window.schema_editor_translations['schema_template_must_be_set']);
-          template_group.addClass("has-error");
-          has_error = true;
-        } else {
-          template_help.text("");
-          template_group.removeClass("has-error");
-        }
       }
       var required = $('#schema-editor-object__' + path.join('__') + '-required-input').prop('checked');
       var has_note = $('#schema-editor-object__' + path.join('__') + '-generic-note-checkbox').prop('checked');
@@ -833,15 +833,30 @@ $(function() {
       updateSpecificProperty(path, real_path, schema, property_schema, has_error);
     }
 
-    function updateTemplateObjectProperty(path, real_path, template_id) {
+    function updateTemplateObjectProperty(path, real_path) {
       var has_error = false;
       updateGenericProperty(path, real_path);
       var schema = JSON.parse(input_schema.text());
       var property_schema = schema['properties'][real_path[real_path.length-1]];
+
       property_schema["type"] = "object";
       property_schema["properties"] = {};
-      var template_id = $('#schema-editor-object__' + path + '-template-id-input').val()
-      property_schema['template'] = template_id;
+
+      var template_input = $('#schema-editor-object__' + path.join('__') + '-template-id-input');
+      var template_group = template_input.closest('.schema-editor-property-settings.schema-editor-template-property-settings');
+      var template_help = template_group.find('.help-block');
+      if($(template_input).is(':visible')) {
+        let new_template_id = $( template_input ).val();
+        if(new_template_id === "" || Number.isNaN(Number.parseInt(new_template_id))) {
+          template_help.text(window.schema_editor_translations['schema_template_must_be_set']);
+          template_group.addClass("has-error");
+          has_error = true;
+        } else {
+          property_schema['template'] = Number.parseInt(new_template_id);
+          template_help.text("");
+          template_group.removeClass("has-error");
+        }
+      }
 
       updateSpecificProperty(path, real_path, schema, property_schema, has_error);
     }
@@ -1113,6 +1128,12 @@ $(function() {
       default_checkbox.prop('checked', false);
     }
     default_checkbox.on('change', updateProperty.bind(path));
+
+    var template_id_input = node.find('.schema-editor-generic-property-template-id-input');
+    template_id_input.attr('id', 'schema-editor-object__' + path.join('__') + '-template-id-input');
+    template_id_input.val(type);
+    template_id_input.selectpicker();
+    template_id_input.on('change', updateProperty.bind(path));
 
     return node;
   }